// frontend/src/app/routers/ventas/formas-pagos/models/formasPagos.ts
export interface FormasPagos {
    forpag_nid?: number;
    cia_nid?: number;
    forpag_vcodigo?: string;
    forpag_vdescripcion?: string;
    forpag_ndprocentajeadicional?: number;
    forpag_nsts?: number;
}

export interface FormasPagosSts {
    forpag_nid?: number;
    forpag_nsts?: number;
}